refactor: move single type export to index
