--- lib/shortlink/types.ts
@@ -1,9 +1,0 @@
-/**
- * Copyright (c) 2023 Quadient Group AG
- * SPDX-License-Identifier: MIT
- */
-
-export interface ShortlinkResponse {
-  url: string
-  shortlink?: string
-}
